Removed jquery dependencies
