+++ test.html
@@ -28,7 +28,16 @@
 	
 	</style>
 
+	<script src="scrollify.js"></script>
+
 	<script>
-	
+	document.addEventListener("DOMContentLoaded", (event) =>
+	{ 
+        scrollify.init(
+		{
+			section : "section",
+		}
+		);
+    });
 	</script>
 </html>
